pyproject.toml
[tool.poetry]
name = "taek-productivity"
version = "0.1.0"
description = ""
authors = ["hoetaek <hoetaek@student.snue.ac.kr>"]
readme = "README.md"
packages = [{include = "taek_productivity"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
